+++ R/brokenstick-class.R
@@ -35,6 +35,8 @@
 #'    `mod` elements.
 #'    The light object can be used to predict broken stick estimates for
 #'    new data, but does not disclose the training data and is small.}
+#'    \item{`hide`}{Should the output for boundary knots be hidden? Can
+#'    be `"right"`, `"left"`, `"both"` or `"none"`. The default is `"right"`.}
 #'    \item{`sigma2j`}{Numeric vector with estimates of the residual variance per
 #'    group. Only used by method `"kr"`.}
 #'    \item{`data`}{The training data used to fit the model.}
@@ -68,6 +70,7 @@ new_brokenstick <- function(call = match.call(),
                             sigma2j = numeric(0),
                             sigma2 = numeric(0),
                             light = FALSE,
+                            hide = NA_character_,
                             sample = numeric(0),
                             data = numeric(0),
                             imp = numeric(0),
@@ -84,7 +87,8 @@ new_brokenstick <- function(call = match.call(),
     omega = omega,
     sigma2 = sigma2,
     sample = sample,
-    light = light
+    light = light,
+    hide = hide
   )
   if (!light) {
     result$sigma2j <- sigma2j